+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Proxy access restriction by IP address/subnet
 - Persist the access log in the diskcache (optional)
+- Show gemtext links at the top of the page to cache the converted webpage
+  (for a certain period of time or forever)
 
 ## [1.3.1] - 2024-03-03